+++ src/TrelloDotNet/Model/Options/GetBoardOptions/GetBoardOptions.cs
@@ -51,6 +51,11 @@ namespace TrelloDotNet.Model.Options.GetBoardOptions
         public bool IncludePluginData { get; set; }
 
         /// <summary>
+        /// What types of boards are returned (Default: All)
+        /// </summary> 
+        public GetBoardOptionsFilter Filter { get; set; } = GetBoardOptionsFilter.All;
+
+        /// <summary>
         /// Additional Parameters not supported out-of-the-box
         /// </summary>
         public List<QueryParameter> AdditionalParameters { get; set; } = new List<QueryParameter>();
@@ -82,6 +87,7 @@ namespace TrelloDotNet.Model.Options.GetBoardOptions
             }
 
             parameters.Add(new QueryParameter("lists", IncludeLists.GetJsonPropertyName()));
+            parameters.Add(new QueryParameter("filter", Filter.GetJsonPropertyName()));
             parameters.Add(new QueryParameter("cards", IncludeCards.GetJsonPropertyName()));
             parameters.Add(new QueryParameter("labels", IncludeLabels ? "all" : "none"));
             parameters.Add(new QueryParameter("pluginData", IncludePluginData));